Cleaned up asub C files

+++ bcmApp/src/asubRWF.c
@@ -1,21 +1,19 @@
-#include <stdio.h>
-#include <stdint.h>
-
 #include <registryFunction.h>
 #include <epicsExport.h>
 #include <aSubRecord.h>
 
 static long calcRWF16(aSubRecord *pasub) {
     /*
-     * Convert a list of 32 bit values into a list (twice the size) of unsigned 16 bit values.
-     *
-     * The lower 2 bytes come first. ex:
-     * 0xcdcdabab should become (0xabab, 0xcdcd) in that order.
+     * Calculate an unsiged 16 bit array from a signed array.
+     * The raw 16 bit stream data is encoded unsigned but the asyn port driver converts it to signed.
+     * This routine eliminates the sign bit with the side effet of halving the sampled value.
      */
-    int i;
+    short i;
     pasub->pact = 1;
 
     for (i = 0; i < pasub->noa; i++) {
+        /* the casts are necessary to first convert from void * and then to do our conversion from signed to
+         * unsigned and back to the signed SHORT the record is expecting */
         ((short *)pasub->vala)[i] = (short)(((unsigned short *)pasub->a)[i] >> 1);
     }
